write: cast video duration to float before round

+++ lib/Db/TimelineWrite.php
@@ -123,7 +123,7 @@ class TimelineWrite
         // Video parameters
         $videoDuration = 0;
         if ($isvideo) {
-            $videoDuration = round($exif['Duration'] ?? $exif['TrackDuration'] ?? 0);
+            $videoDuration = round((float) ($exif['Duration'] ?? $exif['TrackDuration'] ?? 0));
         }
 
         // Clean up EXIF to keep only useful metadata
